requirements.txt
omnidimension
python-dotenv
mysql-connector-python
requests
